make tests isomorphic to source code
